+++ scripts/build.js
@@ -1,8 +1,8 @@
-const startGanache = require('./start-ganache')
-const buildContracts = require('./build-contracts')
-const deployContracts = require('./deploy-contracts')
-const startIpfs = require('./start-ipfs')
-const startTestServer = require('./start-test-server')
+const startGanache = require('./helpers/start-ganache')
+const buildContracts = require('./helpers/build-contracts')
+const deployContracts = require('./helpers/deploy-contracts')
+const startIpfs = require('./helpers/start-ipfs')
+const startTestServer = require('./helpers/start-test-server')
 const watch = require('node-watch')
 const webpack = require('webpack')
 const webpackConfig = require('../webpack.config.js')
