+++ scripts/helpers/build-contracts.js
