+++ scripts/helpers/start-ipfs.js
