+++ scripts/helpers/test-javascript.js
