+++ scripts/test-contracts.js
@@ -1,16 +1,34 @@
-const { spawn } = require('child_process')
+const startGanache = require('./helpers/start-ganache')
+const testContracts = require('./helpers/test-contracts')
+const watch = require('node-watch')
 
-const testContracts = () => {
-  return new Promise((resolve, reject) => {
-    const truffleTest = spawn('../node_modules/.bin/truffle', ['test'], { cwd: './contracts' })
-    truffleTest.stdout.pipe(process.stdout)
-    truffleTest.stderr.on('data', data => {
-      reject(String(data))
-    })
-    truffleTest.on('exit', code => {
-      resolve()
-    })
+// Simple enqueueing system to prevent interrupting a test. Rerunning in the middle of a test causes issues.
+let isRunning = false
+let isEnqueued = false
+const runTests = async () => {
+  if (!isRunning) {
+    isRunning = true
+    await testContracts()
+    isRunning = false
+    if (isEnqueued) {
+      isEnqueued = false
+      runTests()
+    }
+  } else {
+    isEnqueued = true
+  }
+}
+
+const start = async () => {
+  await startGanache()
+
+  runTests()
+
+  // watch contracts
+  watch('./contracts/contracts', { recursive: true }, async (evt, name) => {
+    console.log('%s changed.', name)
+    runTests()
   })
 }
 
-module.exports = testContracts
+start()
